added test coverage to all componenets and renamed Nav to ParentNav for more clarity on use
