+++ src/components/pages/ParentDash/RenderParentDash.js
@@ -3,16 +3,16 @@ import { Route } from 'react-router-dom';
 import './dash.css';
 
 import { Layout } from 'antd';
-import { Nav, DashHome, ChildSignup } from './components';
+import { ParentNav, DashHome, ChildSignup } from './components';
 
 // ParentDash component that contains a nav bar and routes to the various components
-const RenderParentDash = props => {
+const RenderParentDash = () => {
   return (
     <>
       <Layout style={{ background: 'white', height: '100vh' }}>
-        <Nav />
-        <Route exact path={props.match.path} component={DashHome} />
-        <Route path={`${props.match.path}/add`} component={ChildSignup} />
+        <ParentNav />
+        <Route exact path={'/login'} component={DashHome} />
+        <Route path={`/login/add`} component={ChildSignup} />
       </Layout>
     </>
   );
