+++ src/components/pages/ParentDash/components/LinkButton.js
@@ -15,7 +15,6 @@ const LinkButton = props => {
   } = props;
   return (
     <button
-      className="AddCard"
       {...rest}
       onClick={() => {
         history.push(to);
