+++ src/components/pages/ParentDash/components/index.js
@@ -1,4 +1,4 @@
-export { default as Nav } from './Nav.js';
+export { default as ParentNav } from './ParentNav.js';
 export { default as LinkButton } from './LinkButton.js';
 export { default as ChildSignup } from './ChildSignup.js';
 export { default as DashHome } from './DashHome.js';
